Temporarily building with -fno-stack-protector when platform=solo5
Right now, rumprun's libc is built with stack protection. It uses thread local
storage (TLS) to store the canary. This is problematic when running on top of
solo5 as it doesn't allow the TLS register (%fs in x84/64) to be set. So, in
this change we temporarily disable stack protection (build with
-fno-stack-protector) until solo5/solo5 adds a solo5 API to set the TLS
register.

Signed-off-by: Ricardo Koller <kollerr@us.ibm.com>